🔥 [style] : NoteSplitViewController 설계를 위해 삭제 후 코드 분리
